move commands into their own directory
