+++ src/inspect_ai/experimental/_human_agent/commands/submit.py
@@ -1,0 +1,57 @@
+from argparse import Namespace
+from pathlib import Path
+from typing import Awaitable, Callable
+
+from pydantic import JsonValue
+
+from ..state import HumanAgentState
+from .command import HumanAgentCommand, call_human_agent
+
+
+class SubmitCommand(HumanAgentCommand):
+    @property
+    def name(self) -> str:
+        return "submit"
+
+    @property
+    def description(self) -> str:
+        return "Submit your answer for the task."
+
+    @property
+    def cli_args(self) -> list[HumanAgentCommand.CLIArg]:
+        return [
+            HumanAgentCommand.CLIArg(
+                name="answer",
+                description="Answer to submit for scoring (optional, not required for all tasks)",
+            )
+        ]
+
+    def cli(self, args: Namespace) -> None:
+        # read cli args
+        call_args = vars(args)
+
+        # collect session logs if they exist
+        sessions_dir = Path("/var/tmp/user-sessions")
+        if sessions_dir.exists() and sessions_dir.is_dir():
+            session_logs: dict[str, str] = {}
+            for file in sessions_dir.iterdir():
+                if file.is_file():
+                    try:
+                        with open(file, "r") as f:
+                            session_logs[file.name] = f.read()
+                    except Exception as e:
+                        print(f"Error reading file {file.name}: {e}")
+                        continue
+            call_args["session_logs"] = session_logs
+
+        call_human_agent("submit", **call_args)
+
+    def service(self, state: HumanAgentState) -> Callable[..., Awaitable[JsonValue]]:
+        async def submit(
+            answer: str | None, session_logs: dict[str, str] | None = None
+        ) -> None:
+            state.running = False
+            state.answer = answer
+            state.session_logs = session_logs
+
+        return submit
